+++ packages/api/src/services/create_user.ts
@@ -1,29 +1,22 @@
 import { EntityManager } from 'typeorm'
+import { Filter } from '../entity/filter'
 import { GroupMembership } from '../entity/groups/group_membership'
 import { Invite } from '../entity/groups/invite'
 import { Profile } from '../entity/profile'
 import { StatusType, User } from '../entity/user'
+import { env } from '../env'
 import { SignupErrorCode } from '../generated/graphql'
+import { createPubSubClient } from '../pubsub'
 import { authTrx, entityManager, getRepository } from '../repository'
 import { userRepository } from '../repository/user'
 import { AuthProvider } from '../routers/auth/auth_types'
+import { analytics } from '../utils/analytics'
+import { IntercomClient } from '../utils/intercom'
 import { logger } from '../utils/logger'
 import { validateUsername } from '../utils/usernamePolicy'
+import { addPopularReadsForNewUser } from './popular_reads'
 import { sendConfirmationEmail } from './send_emails'
-import { Filter } from '../entity/filter'
-import { analytics } from '../utils/analytics'
-import { env } from '../env'
 
-const TOP_USERS = [
-  'jacksonh',
-  'nat',
-  'luis',
-  'satindar',
-  'malandrina',
-  'patrick',
-  'alexgutjahr',
-  'hongbowu',
-]
 export const MAX_RECORDS_LIMIT = 1000
 
 export const createUser = async (input: {
@@ -111,6 +104,7 @@ export const createUser = async (input: {
       }
 
       await createDefaultFiltersForUser(t)(user.id)
+      await addPopularReadsForNewUser(user.id, t)
 
       return [user, profile]
     }
@@ -121,6 +115,26 @@ export const createUser = async (input: {
       return Promise.reject({ errorCode: SignupErrorCode.InvalidEmail })
     }
   }
+
+  const customAttributes: { source_user_id: string } = {
+    source_user_id: user.sourceUserId,
+  }
+  await IntercomClient?.contacts.createUser({
+    email: user.email,
+    externalId: user.id,
+    name: user.name,
+    avatar: profile.pictureUrl || undefined,
+    customAttributes: customAttributes,
+    signedUpAt: Math.floor(Date.now() / 1000),
+  })
+
+  const pubsubClient = createPubSubClient()
+  await pubsubClient.userCreated(
+    user.id,
+    user.email,
+    user.name,
+    profile.username
+  )
 
   analytics.track({
     userId: user.id,
